Test travis

+++ install_build_run_test_ci.sh
@@ -1,4 +1,8 @@
-yum install -y gcc gcc-c++ libstdc++-devel boost-devel make cmake git wget openssl-libs openssl-devel cyrus-sasl-devel cyrus-sasl-lib
+if [ -f /etc/redhat-release ]; then
+    yum install -y gcc gcc-c++ libstdc++-devel boost-devel make cmake git wget openssl-libs openssl-devel cyrus-sasl-devel cyrus-sasl-lib
+else
+    sudo apt-get install gcc g++ libboost-dev cmake zlib1g-dev libssl1.0.0 libsasl2-2 libssl-dev libsasl2-dev
+fi
 
 git clone https://github.com/edenhill/librdkafka.git
 cd librdkafka
